pyproject.toml
[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["prometheus*"]

[project]
name = "prometheus-v2"
version = "0.1.0"
description = "Prometheus v2 - Quantitative Trading System"
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "Prometheus Team" }
]

dependencies = [
  "numpy>=1.24.0",
  "pandas>=2.0.0",
  "pydantic>=2.0.0",
  "pydantic-settings>=2.0.0",
  "psycopg2-binary>=2.9.0",
  "sqlalchemy>=2.0.0",
  "alembic>=1.12.0",
  "python-dateutil>=2.8.0",
  "pyyaml>=6.0",
  "python-dotenv>=1.0.0",
  "requests>=2.31.0",
  "fastapi>=0.104.0",
  "uvicorn[standard]>=0.24.0",
  "ib_insync>=0.9.86",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4.0",
  "pytest-cov>=4.1.0",
  "pytest-asyncio>=0.21.0",
  "mypy>=1.5.0",
  "ruff>=0.6.0",
  "pydocstyle>=6.3.0",
  "types-PyYAML",
  "types-python-dateutil",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers --tb=short -m 'not integration'"
markers = [
  "integration: marks tests that require external services like databases",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict = true

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
  "E",   # pycodestyle errors
  "W",   # pycodestyle warnings
  "F",   # pyflakes
  "I",   # isort
  "N",   # pep8-naming
  "UP",  # pyupgrade
]
ignore = []

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = ["D", "ANN"]

[tool.coverage.run]
source = ["prometheus"]
omit = ["tests/*", "prometheus/scripts/*"]

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "def __repr__",
  "raise AssertionError",
  "raise NotImplementedError",
  "if __name__ == .__main__.",
  "if TYPE_CHECKING:",
]
